:up: split git and platform for better readability
