+++ functions/platform.py
@@ -88,62 +88,6 @@ def findFolder(path : str, foldername : str)  -> str:
         if foldername in dirs :
             return relpath( str('/'.join([cur,foldername])) )
 
-#get git Rep out of a path
-def getGitRepo(path : str):
-    import os.path
-    path = os.path.realpath(path)
-    repo = None
-    try:
-        from git import Repo
-        try :
-            repo = Repo(path)
-        except:
-            parent_path = None
-            is_git = False
-            while not is_git : 
-                parent_path = os.path.abspath(os.path.join(path, os.pardir))
-                is_git = ".git" in getFolders(parent_path)
-            if not is_git : 
-                raise Exception
-            repo = Repo(parent_path)
-    except:
-        return None
-    finally:
-        return repo
-
-
-# look for path in tree
-def _gitRecursiveLookForPath(path, trees, found = None) :
-    if  not isinstance(found, list) :
-        found = []
-    import os.path
-    try:
-        for tree in trees :
-            try :
-                if str(os.path.basename(path)) in str(os.path.basename(tree.path))  :
-                    found = [str(p.path) for p in tree.trees]
-                else:
-                    _gitRecursiveLookForPath(path, tree.trees, found)
-            except:
-                continue
-    except:
-        pass
-    return found
-    
-
-
-# get names of git tracked folders in path
-def getGitFolders(path : str) -> list :
-    dirs = []
-    try:
-        import os.path 
-        repo = getGitRepo(path)
-        dirs = _gitRecursiveLookForPath(path, repo.tree())
-    except:
-        raise
-    finally:
-        return dirs
-
 # get names of subfolders in folder
 def getFolders(path : str)  -> list:
     from os import walk
